+++ precompiled_headers/pch.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <iostream>
+#include <time.h>
+#include <string>
+#include <stdlib.h>
